+++ src/aux-server/aux-web/shared/scene/Arrow3D.ts
@@ -8,10 +8,10 @@ import {
 import { AuxBot3D } from './AuxBot3D';
 import { DimensionGroup3D } from './DimensionGroup3D';
 import { BuilderGroup3D } from './BuilderGroup3D';
-import { disposeMaterial } from './SceneUtils';
+import { disposeMaterial, buildSRGBColor } from './SceneUtils';
 
 export class Arrow3D extends Object3D {
-    public static DefaultColor: Color = new Color(1, 1, 1);
+    public static DefaultColor: Color = buildSRGBColor(1, 1, 1);
     public static DefaultHeadWidth = 0.15;
     public static DefaultHeadLength = 0.3;