+++ src/aux-server/aux-web/shared/scene/decorators/UserMeshDecorator.ts
@@ -11,7 +11,7 @@ import {
     calculateBooleanTagValue,
     DEFAULT_PLAYER_USER_COLOR,
 } from '@casual-simulation/aux-common';
-import { disposeMesh, createUserCone } from '../SceneUtils';
+import { disposeMesh, createUserCone, buildSRGBColor } from '../SceneUtils';
 import { AuxBot3DDecoratorBase } from '../AuxBot3DDecorator';
 import { AuxBot3D } from '../AuxBot3D';
 import { IMeshDecorator } from './IMeshDecorator';
@@ -95,7 +95,7 @@ export class UserMeshDecorator extends AuxBot3DDecoratorBase
 
         const color = DEFAULT_PLAYER_USER_COLOR;
         const material: MeshToonMaterial = <MeshToonMaterial>this.mesh.material;
-        material.color.set(new Color(color));
+        material.color.set(buildSRGBColor(color));
 
         this.container.visible = this._isActive(calc);
     }